Update amazon-kinesis-client to version 1.4.0
The module ships with version 1.2.0 of the kinesis client
which doesn't support aggregation. I've updated some of the
supporting libraries too, though I'm not 100% sure if the
list is correct.

This configuration works for me on EC2 linux and Mac OS X.
Updating the kinesis client beyond 1.4.0 has issues on
EC2 linux as the default java version is 1.7 and it appears
the client has been compiled against 1.8